+++ tests/unit/Converter/Setter/AddPrefixTest.php
@@ -3,18 +3,17 @@ declare(strict_types=1);
 
 namespace Smile\GdprDump\Tests\Unit\Converter\Setter;
 
-use InvalidArgumentException;
-use Smile\GdprDump\Converter\Setter\SetPrefix;
+use Smile\GdprDump\Converter\Setter\AddPrefix;
 use Smile\GdprDump\Tests\Unit\TestCase;
 
-class SetPrefixTest extends TestCase
+class AddPrefixTest extends TestCase
 {
     /**
      * Test the converter.
      */
     public function testConverter()
     {
-        $converter = new SetPrefix(['prefix' => 'test_']);
+        $converter = new AddPrefix(['prefix' => 'test_']);
 
         $value = $converter->convert('value');
         $this->assertSame('test_value', $value);
@@ -23,10 +22,10 @@ class SetPrefixTest extends TestCase
     /**
      * Assert that an exception is thrown when the prefix is not set.
      *
-     * @expectedException InvalidArgumentException
+     * @expectedException \InvalidArgumentException
      */
     public function testPrefixNotSet()
     {
-        new SetPrefix();
+        new AddPrefix();
     }
 }